[Server] Now user's profile returns all user data
It returns their avgRating, when they've paid, when they've earned, when
they rated others, when others rated them, and all of their requests,
both open and closed.

Please note, this is expensive and will hopefully only be used to
display a profile.  We can have other endpoints for other uses.