[Client] Added FB login button and changed default hostname
Hostname now defaults to our server that Pramono set up.

I added the FB login button to a login screen that launches when the app
first runs.  I can also detect if a user is currently logged in and skip
prompting the user for logging in.  The FB login MUST be tied to the
backend somehow, which I'll look into tomorrow.

There is now a log out button in the overflow menu, but we should
eventually more this to a navigation drawer or something.