Login with Facebook now works
The client logs in with FB, and the sever is alerted to what FB account
is being used.  The server then returns the Taskr profile ID of the
user.

The "authentication" isn't actually secure because all the client has to
know is the email of the FB account of the user to pretend to be another
user.  We can fix that if we want, but it should function just fine for
the demo as-is.